+++ changelog.md
@@ -4,6 +4,8 @@
 
 - Fixed (hopefully) null deref / softlock caused by Barter shop trying to process despawned items
 - Fixed (hopefully) null deref caused by pooled projectiles being duplicated and bypassing the pooling system
+- Fixed bad trail rendering on Grandmaster's projectiles
+- Fixed several projectile trails rendering above the projectiles themselves
 
 ## 1.27.0 (2025-08-17)